+++ spec/dummy_app/db/migrate/20110208155312_set_up_test_tables.rb
@@ -377,6 +377,13 @@ class SetUpTestTables < ::ActiveRecord::Migration::Current
       t.string :name
       t.text :supplier
     end
+
+    create_table :comments do |t|
+      t.text :body
+      t.string :author_name
+      t.boolean :flagged
+      t.timestamps
+    end
   end
 
   def down